+++ CHANGELOG.md
@@ -58,3 +58,8 @@
 * Documentation completed.
 * Template updates (USDD - Russia).
 * Various fixes and updates.
+
+#### 2024.10.09 (**0.7.2**):
+* Additions: processing and display in preferences.
+* Changes to the preferences interface.
+* Fixes: unfold and metal thickness.